+++ plugins/woocommerce-admin/packages/eslint-plugin/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0
+
+-   Updated `@wordpress/eslint-plugin` dependency to latest version.
+    -   `jsdoc` linting is configured to allow typescript style jsdocs.
+
 # 1.0.0
 
 -   Released package